feat: 이미지 변환 기능 mainpage로 이동 & 해당 변경사항에 맞추어 내부 로직 변경
