+++ src/utils/index.ts
@@ -2,3 +2,4 @@ export { default as formatPhoneNumber } from './formatPhoneNumber';
 export { default as getUserInfoList } from './getUserInfoList';
 export { default as getNextTheme } from './getNextTheme';
 export { default as getPrevTheme } from './getPrevTheme';
+export { default as postConvertedImage } from './postConvertedImage';
